pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lectionary-engines"
version = "0.1.0"
description = "Biblical interpretation CLI using three hermeneutical frameworks"
authors = [{name = "Chris Harrison"}]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "click>=8.1.0",
    "anthropic>=0.40.0",
    "python-dotenv>=1.0.0",
    "rich>=13.7.0",
    "pydantic>=2.5.0",
    "requests>=2.31.0",
    "markdown>=3.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.12.0",
    "ruff>=0.1.8",
]

[project.scripts]
lectionary = "lectionary_engines.cli:main"

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
target-version = "py310"
